move classes to misc.py
